+++ docs/tasks/artifacts/sv_corpus_grad/dist_list_braces/repro/c1_constraint_no_dist.sv
@@ -1,0 +1,4 @@
+class C;
+  rand int x;
+  constraint c1 { x > 100; x < 300; }
+endclass