+++ docs/tasks/artifacts/sv_corpus_grad/dist_list_braces/repro/d2_dist_range_eq.sv
@@ -1,0 +1,4 @@
+class C;
+  rand int x;
+  constraint c1 { x dist { [100:102] := 1, 200 := 2, 300 := 5}; }
+endclass