+++ docs/tasks/artifacts/sv_corpus_grad/dist_list_braces/repro/d4_dist_unweighted_soft.sv
@@ -1,0 +1,4 @@
+class C;
+  rand int x;
+  constraint B3 { soft x dist {5, 8}; }
+endclass