+++ docs/tasks/artifacts/sv_corpus_grad/dist_list_braces/repro/min_dist.sv
@@ -1,0 +1,4 @@
+class a;
+  rand int b;
+  constraint c { b dist {3 := 1, 10 := 2}; }
+endclass